add sweetAlert

+++ client/src/components/Undeposited.vue
@@ -13,6 +13,7 @@
 </template>
 
 <script>
+import Swal from 'sweetalert2'
 export default {
   name: 'Undeposited',
   props: ['waste'],
@@ -31,7 +32,24 @@ export default {
 
     deleteItem () {
       const WasteId = this.waste.WasteId
-      this.$store.dispatch('deleteWaste', { WasteId })
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.$store.dispatch('deleteWaste', { WasteId })
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+        }
+      })
     },
 
     addItem () {
